requirements.txt
certifi==2017.4.17
chardet==3.0.4
Django==1.11.3
django-bootstrap-toolkit==2.15.0
django-forms-bootstrap==3.1.0
django-markup-deprecated==0.0.3
idna==2.5
Markdown==2.6.8
mistune==0.7.4
MySQL-python==1.2.5
olefile==0.44
Pillow==4.2.1
Pygments==2.2.0
pytz==2017.2
qiniu==7.1.5
requests==2.18.1
urllib3==1.21.1
